Add missing dependency on PCL headers
When I try to build this package on jazzy with rosdep, the PCL headers
are missing. This package uses `find_package(PCL)` and includes it in
the headers. Therefore I think it would be best to add it to the
`build_depend` and `build_export_depend` tags of the `package.xml`
files.

In the buildfarm this is technically not needed because the PCL headers
are a `build_export_depend` of `pcl_conversions`, but rosdep ignores
this dependency so it misses the PCL headers.

Signed-off-by: Ramon Wijnands <ramon.wijnands@nobleo.nl>